Reorganize callbacks; Enable best effort for sample suggestions; Write initial test
