+++ src/blop/callbacks/__init__.py
@@ -1,3 +1,3 @@
-from .optimization import BestEffortOptimizationCallback
+from .best_effort import BestEffortOptimizationCallback
 
 __all__ = ["BestEffortOptimizationCallback"]
